Use counter for tracking container states
Container state counts are used for reporting in the `/info` endpoint.
Currently when `/info` is called, each container is iterated over and
the containers 'StateString()' is called. This is not very efficient
with lots of containers, and is also racey since `StateString()` is not
using a mutex and the mutex is not otherwise locked.

We could just lock the container mutex, but this is proven to be
problematic since there are frequent deadlock scenarios and we should
always have the `/info` endpoint available since this endpoint is used
to get general information about the docker host.

Really, these metrics on `/info` should be deprecated. But until then,
we can just keep a running tally in memory for each of the reported
states.

Signed-off-by: Brian Goff <cpuguy83@gmail.com>
(cherry picked from commit e4c03623c2fcd3013343d246e3432409850c8c37)
Signed-off-by: Kir Kolyshkin <kolyshkin@gmail.com>

Conflicts:
 - daemon/delete.go: different imports
 - daemon/info.go: commit 62563e1707f0 missing from upstream
 - vendor/github.com/docker/go-metrics/README.md: newer vndr